refactor: fe - cavnas hook 분리 #82
