+++ frontend/src/pages/workspace/whiteboard-canvas/index.tsx
@@ -1,43 +1,12 @@
-import { useEffect, useRef } from 'react';
 import { fabric } from 'fabric';
-import { drawingGird } from '@utils/fabric.utils';
+import useCanvas from './useCanvas';
 
 function WhiteboardCanvas() {
-	const canvas = useRef<fabric.Canvas | null>(null);
-	useEffect(() => {
-		canvas.current = initCanvas();
-		return () => {
-			if (canvas.current) {
-				canvas.current.dispose();
-				canvas.current = null;
-			}
-		};
-	}, []);
-
-	const initCanvas = () => {
-		const grid = 50;
-		const canvasWidth = 2000;
-		const canvasHeight = 900;
-
-		const fabricCanvas = new fabric.Canvas('canvas', {
-			height: canvasHeight,
-			width: canvasWidth,
-			backgroundColor: '#f1f1f1',
-		});
-
-		drawingGird(fabricCanvas, canvasWidth, canvasHeight, grid);
-
-		// 이벤트 추가 예시
-		fabricCanvas.on('object:added', (e) => {
-			console.log(e);
-		});
-
-		return fabricCanvas;
-	};
+	const { canvas } = useCanvas();
 
 	// Object 추가 예시
 	const addObj = () => {
-		if (!canvas.current) return;
+		if (!canvas) return;
 		const rect = new fabric.Rect({
 			height: 280,
 			width: 200,
@@ -45,14 +14,14 @@ function WhiteboardCanvas() {
 			left: 100,
 			fill: 'yellow',
 		});
-		canvas.current.add(rect);
+		canvas.add(rect);
 	};
 	// Object 삭제 예시
 	const clearObjects = () => {
-		if (!canvas.current) return;
-		canvas.current.forEachObject((obj) => {
+		if (!canvas) return;
+		canvas.forEachObject((obj) => {
 			// 그리드 제외 하고 삭제
-			if (!(obj instanceof fabric.Line)) canvas.current?.remove(obj);
+			if (!(obj instanceof fabric.Line)) canvas?.remove(obj);
 		});
 	};
 	return (
